chore: updated tests
back at 100%
